Lots of Changes
All Objects in this game now inherit from the LivingEntity class instead of MonoBehavious so you can set their starting health from inside the inspector and they are able to be killed. The player can now die when in range of the Cow for a long enough and it will open the loseUI. I have buttons set up to return to the main meun and to a settings menu although the menu's haven't been implemented. You can hit space to retry the current level. Just load my scene and hit play to see the changes. I need to work with the resolution to get it to feel right something is weird, I'll work on that next.