Edited Cow Script
Added a circle collider for the player to interact with instead and the cow will immediately turn red when the player is in this zone. It looks stupid but I intend to add some coroutines that make the cow lerp in between the normal color and red so it slowly gets angry when the player is nearby, he will turn red at the same speed as the cows attack time.